chore: move enum and types to nestjs common
