+++ packages/common/enums/index.ts
@@ -2,3 +2,4 @@ export * from './request-method.enum';
 export * from './http-status.enum';
 export * from './shutdown-signal.enum';
 export * from './version-type.enum';
+export * from './lifecycle-type.enum';
