--- packages/core/enums/index.ts
@@ -1,1 +1,0 @@
-export * from './lifecycle-type.enum';
