--- packages/core/enums/lifecycle-type.enum.ts
@@ -1,5 +1,0 @@
-export enum LifeCycleType {
-  GUARDS = 'GUARDS',
-  PIPES = 'PIPES',
-  INTERCEPTORS = 'INTERCEPTORS',
-}
